Update XSS.java
To fix the ReflectXSS vulnerability, weneed to sanitize the xss parameter before returning it in the response. One common way to do this is by using HTML escaping and for both steps of the StoredXSS vulnerability, we should sanitize the xss parameter before storing it in a cookie and before displaying its value.